+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Mi 2024 aditoprim ([doi:10.3389/fphar.2024.1378034](https://doi.org/10.3389/fphar.2024.1378034)) - swine PBPK and integrated PBPK/PD against *Streptococcus suis*.
+
 - Add Kuroda 2023 cephalothin ([doi:10.1294/jes.34.111](https://doi.org/10.1294/jes.34.111)) - Thoroughbred horses given intramuscular and intravenous doses.
 
 - Add Zhang 2024 sertraline ([doi:10.1016/j.heliyon.2024.e25231](https://doi.org/10.1016/j.heliyon.2024.e25231)) - Chinese inpatients with psychiatric disorders, aged 11-79 years.